requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
contourpy==1.2.0
cycler==0.12.1
eccv-caption==0.1.0
filelock==3.9.0
fonttools==4.47.2
fsspec==2023.12.2
ftfy==6.1.3
huggingface-hub==0.20.1
idna==3.4
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lit==15.0.7
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.0
nltk==3.8.1
numpy==1.24.1
packaging==23.2
pandas==2.2.0
Pillow==9.3.0
prettytable==3.9.0
pycocotools==2.0.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
SentEval @ file:///home/SimCSE/SentEval
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
transformers==4.36.2
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.4
urllib3==1.26.13
wcwidth==0.2.12
zipp==3.17.0
